Update Commands and its subclasses execute function
Execute function uses old UI class for text-based UI.

This is incompatible with JavaFX UI Component.

Let's updates them to use GuiController to add user and chatbot
dialog boxs.

This ensures that the GUI is appropriately updated after a
command is executed.
