+++ src/main/java/owen/command/UnmarkCommand.java
@@ -1,15 +1,15 @@
-package command;
+package owen.command;
 
-import storage.Storage;
-import task.TaskList;
-import ui.Ui;
+import owen.storage.Storage;
+import owen.task.TaskList;
+import owen.ui.GuiController;
 
 /**
  * Represents a command to unmark a task from the task list.
  */
 public class UnmarkCommand extends Command {
 
-    /** The keyword to identify this command. */
+    /** The keyword to identify this Owen.command. */
     public static final String KEY_WORD = "unmark";
 
     /** The index of the task to be unmarked. */
@@ -25,9 +25,11 @@ public class UnmarkCommand extends Command {
     }
 
     @Override
-    public void execute(Ui ui, Storage storage, TaskList taskList) {
+    public void execute(GuiController guiController, Storage storage, TaskList taskList) {
         taskList.markTaskAsUndone(pendingTaskIndex);
         storage.overwriteTasklistData(taskList.getTaskList());
-        ui.showMessage("The following is now not done: \n" + taskList.getTaskStatus(pendingTaskIndex));
+        guiController.addUserDialog();
+        String response = "The following is now not done: \n" + taskList.getTaskStatus(pendingTaskIndex);
+        guiController.addOwenDialog(response);
     }
 }
